Handled and improved sign-in system. Added-improved proxy routes.

+++ src/components/overlays/Shading/Shading.tsx
@@ -14,18 +14,18 @@ const Shading: React.FC = () => {
   const isDeletePopupVisible = useAppSelector(state => state.accountUi.isDeletePopupVisible);
 
   const closeShading = () => {
-    dispatch(updateIsShadingVisible(false));
+    if (isShadingVisible) dispatch(updateIsShadingVisible(false));
 
-    dispatch(updateIsSignInPopupVisible(false));
-    dispatch(updateIsSignOutPopupVisible(false));
-    dispatch(updateIsDeletePopupVisible(false));
+    if (isSignInPopupVisible) dispatch(updateIsSignInPopupVisible(false));
+    if (isSignOutPopupVisible) dispatch(updateIsSignOutPopupVisible(false));
+    if (isDeletePopupVisible) dispatch(updateIsDeletePopupVisible(false));
   }
 
   const changeAll = (key: string, boolean: boolean) => {
     if (boolean) {
-      if (key !== 'signInPopup') dispatch(updateIsSignInPopupVisible(false));
-      if (key !== 'signOutPopup') dispatch(updateIsSignOutPopupVisible(false));
-      if (key !== 'deletePopup') dispatch(updateIsDeletePopupVisible(false));
+      if (key !== 'signInPopup' && isSignInPopupVisible) dispatch(updateIsSignInPopupVisible(false));
+      if (key !== 'signOutPopup' && isSignOutPopupVisible) dispatch(updateIsSignOutPopupVisible(false));
+      if (key !== 'deletePopup' && isDeletePopupVisible) dispatch(updateIsDeletePopupVisible(false));
 
       if (!isShadingVisible) {
         dispatch(updateIsShadingVisible(true));
